requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
jinja2==3.1.6
python-multipart==0.0.20
sqlalchemy==2.0.31
bcrypt==4.1.2
pyjwt==2.12.1
pydantic==2.5.3
nbformat==5.10.4
python-dotenv==1.0.1
